Abstract AVR ADC
The AnalogInputPin and AnalogInputController classes where added to the
io component of the AVR_Lib library to accommodate an abstraction for
the analog to digital converter in an AVR microcontroller. As there are
multiple channels but only a single converter in each mcu, a centralized
controller class had to be written.

The UART communication component was refactored to make it more readable
and maintainable using the device macros provided by AVR.

A bug in the buffer class was fixed.