Add external interrupt handling
External interrupt handling was added to the AVR_Lib into the new
interrupt component. It contains a config structure in the usual way,
an ExternalInterrupt class which sets the AVR registers accordingly and
the ExternalInterruptHandler, which has to be derived for any type of
handling activity to provide an interrupt service routine.